% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{add_small_legend}
\alias{add_small_legend}
\title{Add small legend to ggplot}
\usage{
add_small_legend(gg_plot, point_size = 0.5, text_size = 3, space_legend = 0.1)
}
\arguments{
\item{gg_plot}{plot of ggplot type}

\item{point_size}{Size of legend pointers}

\item{text_size}{Size of legend text}

\item{space_legend}{Space between legends}
}
\value{
Plot with adjusted legend
}
\description{
\code{add_small_legend} is a function to adjust the size of a plot legend.
Used in scripts: {}
Robust function: TRUE
}
