% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{add_stars}
\alias{add_stars}
\title{Add stars to statistical estimate}
\usage{
add_stars(estimates, p_vals)
}
\arguments{
\item{estimates}{Numeric, vector with statistical estimates}

\item{p_vals}{Numeric, vector with p-values for each element in
\code{estimates}}
}
\value{
Character, vector with estimates and stars
}
\description{
\code{add_stars} is a function that adds stars to statistical estimates for
easy addition in latex tables.
Used in scripts: {}
Robust function: FALSE
}
