% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{alligator_remover}
\alias{alligator_remover}
\title{Remove alligator signs}
\usage{
alligator_remover(col_names = NULL)
}
\arguments{
\item{col_names}{Character string (vector with strings)}
}
\value{
If input is single string a single string is returned. If input is
vector of strings, a character vector is returned.
}
\description{
\code{alligator_remover} vectorised function that removes "<" and ">" from
strings and replaces spaces between words with underscores (_).
}
\details{
Used in scripts: {01_process_tables.R, 03_process_ts.R}
Robust function: TRUE
}
