% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{ann_monthly_ret}
\alias{ann_monthly_ret}
\title{Annualise monthly returns}
\usage{
ann_monthly_ret(x)
}
\arguments{
\item{x}{vector of percentage returns}
}
\value{
vector of annualised returns in percentage
}
\description{
\code{ann_monthly_ret} annualises monthly returns, when the numbers are
expressed in percentages, i.e. between zero and 100.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
