% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{ann_sd}
\alias{ann_sd}
\title{Annulise standard deviation}
\usage{
ann_sd(x, scale = 12)
}
\arguments{
\item{x}{numeric vector}

\item{scale}{time frequency of \code{x}}
}
\value{
vector of annualised standard deviations
}
\description{
\code{ann_sd} annualises the standard deviation based on scale
(monthly = 12, quarterly = 4, biannually = 2, daily = 252)
}
\details{
Used in scripts: {}
Robust function: TRUE
}
