% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_log_return}
\alias{calc_log_return}
\title{Compute logarithmic net returns for a vector}
\usage{
calc_log_return(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Vector of same length as \code{x}
}
\description{
\code{calc_log_return} helper function for \code{compute_return}.
Takes every element of \code{x} except the first and divides by every
element of \code{x} excep the last and takes the log.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
