% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_simple_return}
\alias{calc_simple_return}
\title{Compute simple net returns for a vector
Used in scripts: {}}
\usage{
calc_simple_return(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
Vector of same length as \code{x}
}
\description{
\code{calc_simple_return} helper function for \code{\link{compute_return}}.
Takes every element of \code{x} except the first and divides by every
element of \code{x} excep the last and subtracts one (1).
}
