% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{camel_to_snake}
\alias{camel_to_snake}
\title{Convert camel case words to snake case words}
\usage{
camel_to_snake(x)
}
\arguments{
\item{x}{Character vector of one or more elements.}
}
\value{
Character vector of camel case words.
}
\description{
\code{camel_to_snake} is a function that splits a word by capital letters,
converts all letters to lower case and bind the splitted words together by
an underscore, e.g., SecurityId becomes security_id. It is primarily used
to fix the column names in the raw data downloaded from Børsprosjektet.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
\examples{
\dontrun{
camel_to_snake("SecurityID")

camel_to_snake("SecId", "FooBar")
}
}
