% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{compute_return}
\alias{compute_return}
\title{Compute simple or logarithmic returns}
\usage{
compute_return(x = NULL, type = c("simple", "log"))
}
\arguments{
\item{x}{numeric vector}

\item{type}{either "simple" or "log"}
}
\value{
If \code{x} is a vector of length > 1, returns a vector of equal
length to input vector with returns and code{NA_real_} in first position
}
\description{
\code{compute_return} computes simple or logarithmic returns for a vector
of values. The function uses \code{\link{calc_simple_return}} and
\code{\link{calc_log_return}}.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
