% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{conditional_drop}
\alias{conditional_drop}
\title{Drop columns from data frame if they exist}
\usage{
conditional_drop(data, cols)
}
\arguments{
\item{data}{Data frame}

\item{cols}{Character vector with names of one or more columns to drop}
}
\value{
\code{conditional_drop} returns the data frame without those columns
that do exist.
}
\description{
\code{conditional_drop} removes a set of columns passed by a character vector
if they exist in the data frame. If none exist, the original dataframe is
returned
}
\details{
Used in scripts: {}
Robust function: FALSE
}
