% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{cumret}
\alias{cumret}
\title{Compute the cumulative product of a vector}
\usage{
cumret(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
\code{cumret} returns the last value in the cumulative product vector
}
\description{
\code{cumret} is a function that computes the cumulative product of a numeric
vector \code{x} and returns the last element.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
