% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{divide_cols_by}
\alias{divide_cols_by}
\title{Divide a set of columns by a number}
\usage{
divide_cols_by(data = NULL, cols = NULL, by = 100)
}
\arguments{
\item{data}{Data frame or tibble object}

\item{cols}{Vector of one or more columns in \code{data} to operate on}

\item{by}{Number to divide by. Default is 100}
}
\value{
Data frame
}
\description{
\code{divide_cols_by} takes a \code{data frame} or \code{tibble} object,
a set of column names in a vector, any real number and divides the values
in the cols argument by that number and returns the tbl.
}
\details{
Used in scripts: {}
Robust function: FALSE

TO IMPROVE:
\itemize{
\item Check that data is not NULL, if return NULL
\item Make column names not case sensitive
\item Check which column(s) are and are not in the dataset, throw warning if
any are missing, and operate only on those that are in the dataset
\item Check the classes of the passed columns to ensure they are numeric and
possible to divide by \code{by}. If not attempt to convert to numeric, else
throw a warning that those columns are not operated on
}
}
