% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_na_df}
\alias{filter_na_df}
\title{Filter away NA rows for multiple columns at at time}
\usage{
filter_na_df(data, cols)
}
\arguments{
\item{data}{Data frame or tibble}

\item{cols}{Character vector of one or more column names found in \code{data}}
}
\value{
Filtered data frame
}
\description{
\code{filter_na_df} is a function that takes a vector of variable names
and filters away all NA values for those columns.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
