% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{gen_missing_dates}
\alias{gen_missing_dates}
\title{Extract missing dates in a monthly sequence}
\usage{
gen_missing_dates(x = NULL)
}
\arguments{
\item{x}{Vector of dates}
}
\value{
If multiple dates are missing, it returns a vector of dates. If one
date is missing, it returns a single date and if no dates are missing it
returns NA
}
\description{
\code{gen_missing_dates} takes a sequence of dates and finds missing
month-end dates. It uses the function \code{\link{gen_month_ends}} to
generate a sequence of month-end dates between the first and last date in
\code{x}.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
