% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{gen_month_ends}
\alias{gen_month_ends}
\title{Sequence of monthly end dates}
\usage{
gen_month_ends(sdate = NULL, edate = NULL)
}
\arguments{
\item{sdate}{Start date on "Y-m-d" format}

\item{edate}{End date on "Y-m-d" format}
}
\value{
Sequence of dates
}
\description{
\code{gen_month_ends} takes a start and end date and creates a sequence of
monthly dates, with the last day of each month.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
