% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Opposite of in function}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{Elements to check that are not in \code{table}}

\item{table}{Table to look up values \code{x} in}
}
\value{
Returns the elements of \code{x} not in \code{table}
}
\description{
Opposite of in function
}
\examples{
c(1, 2, 3) \%nin\% c(2, 3)
}
