% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{has_duplicates}
\alias{has_duplicates}
\title{Are there duplicate values in your data?}
\usage{
has_duplicates(data, ...)
}
\arguments{
\item{data}{Dataframe/tibble}

\item{...}{Other arguments}
}
\value{
TRUE if there are duplicate rows, otherwise FALSE
}
\description{
\code{has_duplicates} checks if there are duplicate rows in your dataframe,
unconditional on grouping.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
