% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Interpolate missing values using mean}
\usage{
impute_missing_values(data, max_na = 4)
}
\arguments{
\item{data}{Numeric vector with missing data}

\item{max_na}{Maximum number of NA in a row and still fill in values}
}
\value{
Numeric vector with NA values in groups <= max_na filled in
}
\description{
\code{impute_missing_values} is a function that takes a numeric vector with
missing data and replaces the NAs with the mean of the last value on the
left side and the first non-NA value on the right. It does this only for
repeated values of NA smaller than size \code{max_na}. This means that if
\code{max_na = 4}, the function will only replace missing values if there
are four or fewer missing values between the last non NA on the left and
the first non NA on the right.
}
\examples{
\dontrun{
x <-
c(1, 1, NA, NA, NA, NA, 2, 2, 5, NA, NA, 6, 7, NA, 8, 9, NA,
  NA, NA, NA, NA, 10, 10, 13, NA, NA, 14, NA, NA, NA, NA, NA, NA, NA, 12)

impute_missing_values(data = x)

}
}
