% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lag_cols}
\alias{lag_cols}
\title{Lag multiple columns at at time}
\usage{
lag_cols(data, cols, n = 1)
}
\arguments{
\item{data}{Data frame or tibble}

\item{cols}{Character vector of one or more column names found in \code{data}}

\item{n}{Variable in lag function, number of periods to lag}
}
\value{
Lagged data frame
}
\description{
\code{lag_cols} is a function that takes a vector of variable names
and lags them all.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
