% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{last_day}
\alias{last_day}
\title{Find last day of month}
\usage{
last_day(date)
}
\arguments{
\item{date}{Single date object or vector of dates}
}
\value{
If \code{date} is a single date, a single date is returned. If
\code{date} is a vector of date objects a vector is returned.
}
\description{
\code{last_day} takes a single date or vector of dates and returns the last
date in the month.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
