% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{left_join_with_null}
\alias{left_join_with_null}
\title{Handle a data frame being NULL when merging}
\usage{
left_join_with_null(x = NULL, y = NULL, by = NULL)
}
\arguments{
\item{x}{Data frame}

\item{y}{Data frame}

\item{by}{Character vector of common columns for \code{x} and \code{y}}
}
\value{
If both \code{x} and \code{y} is NULL, then NULL is returned. If
either \code{x} or \code{y} is NULL, the non-NULL object is returned.
Finally, if both \code{x} and \code{y} is not NULL, then the merged data
frame is returned.
}
\description{
\code{left_join_with_null} does not throw an error if \code{x} or \code{y} is
a NULL. In that case it returns the data frame that is not NULL. If both are
NULL, then NULL is returned. This function is based on the
\code{\link{left_join}} function from dplyr package.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
