% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{length_na}
\alias{length_na}
\title{Calculate length of NAs in a vector}
\usage{
length_na(x)
}
\arguments{
\item{x}{Vector}
}
\value{
\code{length_na} returns the number of NAs in a vector.
}
\description{
\code{length_na} is a function that computes how many NA values there are
in a vector
}
\details{
Used in scripts: {}
Robust function: TRUE
}
