% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{look_ncols_excel}
\alias{look_ncols_excel}
\title{Look up the number of columns in an excel sheet}
\usage{
look_ncols_excel(excel_file = NULL, sheet_num = NULL)
}
\arguments{
\item{excel_file}{Complete file name (path) to the excel file}

\item{sheet_num}{Integer value - the sheet number to read to R}
}
\value{
Returns the number of columns detected by \code{read_excel}
}
\description{
\code{look_ncols_excel} is a function that reads in the first two lines of an
excel sheet and returns the number of columns. It is used as a helper
function to set the correct column types when importing all the time series
data downloaded from Lipper. The reason for this is that the first column in
excel output from Lipper contains the row numbers, but when the sheet has no
data, this first column will not be detected by \code{\link{read_excel}}.
Therefore the column types specified will vary depending on the output from
Lipper.
}
\details{
Robust function: FALSE
}
