% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{path_delist}
\alias{path_delist}
\title{Delist paths}
\usage{
path_delist(path)
}
\arguments{
\item{path}{List of paths}
}
\value{
Vector of file paths of class \code{fs}
}
\description{
\code{path_delist} is inspired by the functions in the \code{fs} package.
It takes a list of file paths and delist it and converts it to an fs object
}
\details{
Used in scripts: {}
Robust function: FALSE
}
