% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{push_to_month_end}
\alias{push_to_month_end}
\title{Convert date to the last day of the month}
\usage{
push_to_month_end(date)
}
\arguments{
\item{date}{Date on format from Lubridate package}
}
\value{
Returns the date of the last day of the month
}
\description{
\code{push_to_month_end} is a function that returns the date of the last
day of the month for a given date.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
