% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_colnames}
\alias{read_colnames}
\title{Pull the column names from a data frame stored in fst format}
\usage{
read_colnames(filep = NULL)
}
\arguments{
\item{filep}{Character vector of a single file path}
}
\value{
Vector with column names
}
\description{
\code{read_colnames} takes a file path to a fst file, reads in the first
row of the data frame and returns the column names.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
