% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_nrows}
\alias{read_nrows}
\title{Pull the number of rows from a data frame stored in fst format}
\usage{
read_nrows(filep = NULL)
}
\arguments{
\item{filep}{Character vector of a single file path}
}
\value{
Numeric - number of rows
}
\description{
\code{read_nrows} takes a file path to a fst file, reads in the first
column of the data frame and returns the number of rows. It relies on
\code{\link{read_colnames}} to get the first column name of the data frame.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
