% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{standardise_character}
\alias{standardise_character}
\title{Standardise a character string}
\usage{
standardise_character(x)
}
\arguments{
\item{x}{Character vector of one or more strings}
}
\value{
Character vector
}
\description{
\code{standardise_character} takes a character vector and transform all
elements to lowercase, removes trailing, leading and superflous whitespaces
and replaces all single spaces with an underscore.
}
\details{
Used in scripts: {}
Robust function: FALSE
}
