% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_remove_after}
\alias{str_remove_after}
\title{Remove everything after the first separator in a string}
\usage{
str_remove_after(string, sep)
}
\arguments{
\item{string}{Input vector.}

\item{sep}{Where to split the string.}
}
\value{
Character vector with only the parts appearing before the separator.
}
\description{
\code{str_remove_after} is a (vectorised) function that removes everything
that appears after a chosen separator in a text string. This is primarily
used to remove characters in security names that appear after the separator
in order to make it easier to match companies by name.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
