% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_remove_decimals}
\alias{str_remove_decimals}
\title{Remove decimals from large numbers in string}
\usage{
str_remove_decimals(string, pattern = NULL, n_digits = 2)
}
\arguments{
\item{string}{Character string}

\item{pattern}{Number pattern to look for}

\item{n_digits}{Indicating how large the number must be before removing
decimals. n_digits = 2 will remove decimals for numbers greater than or
equal to 10, n_digits = 3 for numbers greater than or equal to 100 etc.}
}
\value{
Character string with decimals removed for large numbers
}
\description{
\code{str_remove_decimals} is a function that for a character vector where
each element is a single word or number identifies all decimal numbers
(detected by a period .) and if the preceeding number of digits is larger
than one, i.e., the number in the text is >=10, then the decimals are removed
from the number. The intended use is to clean output to latex tables such
that decimals for large numbers are removed. It is important that the numbers
are stored in a single element in the string vector.
}
\details{
Used in scripts: {}
Robust function: TRUE
}
\examples{

\dontrun{
str_remove_decimals("Hello, world 1,000.000")
str_remove_decimals("1.000")
str_remove_decimals("10.000")
str_remove_decimals("1,000.00")
str_remove_decimals(c("Hello, world", "1,000.00"))
}
}
