% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_wss_latex}
\alias{theme_wss_latex}
\title{Custom Latex theme}
\usage{
theme_wss_latex(base_size = 10, base_family = "")
}
\arguments{
\item{base_size}{The standard size for text in the figure}

\item{base_family}{Left open}
}
\value{
Returns the same figure, only with a new theme
}
\description{
\code{theme_wss_latex} is a custom theme for figures that will be used in
latex files, i.e. research papers. The theme is quite minimalistic and clean
as figures used in research should be. The theme is prone to adjustments in
the future
}
