% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{unlist_dates}
\alias{unlist_dates}
\title{Convert dates in list (data frame or list) to vector of dates}
\usage{
unlist_dates(x)
}
\arguments{
\item{x}{Data frame or list of dates}
}
\value{
Vector of dates on "\%Y-\%m-\%d" format
}
\description{
\code{unlist_dates} converts dates stored in data frames or lists into
a vector while retaining the date format. It also unnames the dates and
uses the \code{lubridate} package to convert the dates to \code{ymd} format
}
\details{
Used in scripts: {}
Robust function: FALSE
}
